+++ packages/engine-svg/CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Updated engine to new `registerFrameCallback` and `registerDoneCallback` setup
-- Breaking: `mechanic` is now destructured into the design function directly, so you can access `drawLoop`, `frame`,`done` directly in the function arguments instead of needing to destructure them from `mechanic`. `setState` and `state` remain on the `mechanic` object.
+- `done` and `frame` are passed into the design function directly now. Destructuring them from `mechanic` is deprecated.
 
 ## 2.0.0-beta.0 - 2022-03-16